apps/sidf_home/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0


streamlit==1.40.2
boto3==1.35.76
pdf2image==1.17.0
python-docx==1.1.2

pillow==10.4.0
python-dotenv==1.0.1
libreoffice_convert==1.0
poppler-utils==0.1.0
PyPDF2==3.0.1
docling-core==2.12.1
docling==2.14.0